+++ client/src/lib.rs
@@ -48,6 +48,9 @@ mod transaction;
 pub use crate::interface::*;
 pub use radicle_registry_core::Balance;
 pub use radicle_registry_runtime::fees::MINIMUM_FEE;
+pub use radicle_registry_runtime::registry::{
+    REGISTER_MEMBER_DEPOSIT, REGISTER_ORG_DEPOSIT, REGISTER_PROJECT_DEPOSIT, REGISTER_USER_DEPOSIT,
+};
 
 pub use backend::{EmulatorControl, EMULATOR_BLOCK_AUTHOR};
 
@@ -283,13 +286,13 @@ impl ClientT for Client {
     async fn get_project(
         &self,
         project_name: ProjectName,
-        org_id: Id,
+        project_domain: ProjectDomain,
     ) -> Result<Option<Project>, Error> {
-        let project_id = (project_name.clone(), org_id.clone());
+        let project_id = (project_name.clone(), project_domain.clone());
         self.fetch_map_value::<registry::store::Projects, _, _>(project_id.clone())
             .await
             .map(|maybe_project| {
-                maybe_project.map(|project| Project::new(project_name, org_id, project))
+                maybe_project.map(|project| Project::new(project_name, project_domain, project))
             })
     }