+++ client/tests/end_to_end.rs
@@ -16,6 +16,7 @@
 //! Test the client against a running node.
 //!
 //! Note that chain state is shared between the test runs.
+//! To avoid AccountUserAssociated errors, use a distinct author for each test.
 
 use serial_test::serial;
 
@@ -28,20 +29,99 @@ async fn register_project() {
     let _ = env_logger::try_init();
     let node_host = url::Host::parse("127.0.0.1").unwrap();
     let client = Client::create_with_executor(node_host).await.unwrap();
-    let (author, _) = key_pair_with_associated_user(&client).await;
+    let author = random_key_pair(&client).await;
 
-    let project_hash = H256::random();
-    let checkpoint_id = submit_ok(
-        &client,
-        &author,
-        message::CreateCheckpoint {
-            project_hash,
-            previous_checkpoint_id: None,
-        },
-    )
-    .await
-    .result
-    .unwrap();
+    for domain in generate_project_domains(&client, &author).await {
+        let project_hash = H256::random();
+        let checkpoint_id = submit_ok(
+            &client,
+            &author,
+            message::CreateCheckpoint {
+                project_hash,
+                previous_checkpoint_id: None,
+            },
+        )
+        .await
+        .result
+        .unwrap();
+
+        let initial_balance = match &domain {
+            ProjectDomain::Org(org_id) => {
+                let org = client.get_org(org_id.clone()).await.unwrap().unwrap();
+                client.free_balance(&org.account_id).await.unwrap()
+            }
+            ProjectDomain::User(user_id) => {
+                let user = client.get_user(user_id.clone()).await.unwrap().unwrap();
+                client.free_balance(&user.account_id).await.unwrap()
+            }
+        };
+
+        let random_fee = random_balance();
+        let message = random_register_project_message(&domain, checkpoint_id);
+        let tx_included = submit_ok_with_fee(&client, &author, message.clone(), random_fee).await;
+
+        let project = client
+            .get_project(message.project_name.clone(), message.project_domain.clone())
+            .await
+            .unwrap()
+            .unwrap();
+        assert_eq!(project.name.clone(), message.project_name.clone());
+        assert_eq!(project.domain.clone(), message.project_domain.clone());
+        assert_eq!(project.current_cp.clone(), checkpoint_id);
+        assert_eq!(project.metadata.clone(), message.metadata.clone());
+
+        assert_eq!(
+            tx_included.events[0],
+            RegistryEvent::ProjectRegistered(
+                message.project_name.clone(),
+                message.project_domain.clone()
+            )
+            .into()
+        );
+
+        let has_project = client
+            .list_projects()
+            .await
+            .unwrap()
+            .iter()
+            .any(|id| *id == (message.project_name.clone(), message.project_domain.clone()));
+        assert!(has_project, "Registered project not found in project list");
+
+        let checkpoint_ = state::Checkpoint {
+            parent: None,
+            hash: project_hash,
+        };
+        let checkpoint = client.get_checkpoint(checkpoint_id).await.unwrap().unwrap();
+        assert_eq!(checkpoint, checkpoint_);
+
+        let (projects, account_id) = match &domain {
+            ProjectDomain::Org(org_id) => {
+                let org = client.get_org(org_id.clone()).await.unwrap().unwrap();
+                (org.projects, org.account_id)
+            }
+            ProjectDomain::User(user_id) => {
+                let user = client.get_user(user_id.clone()).await.unwrap().unwrap();
+                (user.projects, user.account_id)
+            }
+        };
+
+        assert_eq!(projects, vec![project.name]);
+        assert_eq!(
+            client.free_balance(&account_id).await.unwrap(),
+            initial_balance - random_fee,
+            "The tx fee was not charged properly."
+        );
+    }
+}
+
+#[async_std::test]
+#[serial]
+async fn register_member() {
+    let _ = env_logger::try_init();
+    let node_host = url::Host::parse("127.0.0.1").unwrap();
+    let client = Client::create_with_executor(node_host).await.unwrap();
+    let (author, author_id) = key_pair_with_associated_user(&client).await;
+    let (_, user_id) = key_pair_with_associated_user(&client).await;
 
     let org_id = random_id();
     let register_org_message = message::RegisterOrg {
@@ -55,58 +135,41 @@ async fn register_project() {
     let initial_balance = 1000;
     transfer(&client, &author, org.account_id, initial_balance).await;
 
-    let register_project_message = random_register_project_message(org_id.clone(), checkpoint_id);
-    let project_name = register_project_message.project_name.clone();
+    assert_eq!(org.members, vec![author_id.clone()]);
+
+    let register_member_message = message::RegisterMember {
+        org_id: org_id.clone(),
+        user_id: user_id.clone(),
+    };
     let random_fee = random_balance();
-    let tx_included = submit_ok_with_fee(
+    let tx_applied = submit_ok_with_fee(
         &client,
         &author,
-        register_project_message.clone(),
+        register_member_message.clone(),
         random_fee,
     )
     .await;
-    assert_eq!(tx_included.result, Ok(()));
-
-    let project = client
-        .get_project(project_name.clone(), org_id.clone())
-        .await
-        .unwrap()
-        .unwrap();
-    assert_eq!(project.name.clone(), project_name.clone());
-    assert_eq!(project.org_id.clone(), org_id.clone());
-    assert_eq!(
-        project.current_cp.clone(),
-        register_project_message.checkpoint_id
-    );
-    assert_eq!(project.metadata.clone(), register_project_message.metadata);
+    assert_eq!(tx_applied.result, Ok(()));
 
     assert_eq!(
-        tx_included.events[0],
-        RegistryEvent::ProjectRegistered(project_name.clone(), org_id.clone()).into()
+        tx_applied.events[0],
+        RegistryEvent::MemberRegistered(user_id.clone(), org_id.clone()).into()
     );
 
-    let checkpoint = client.get_checkpoint(checkpoint_id).await.unwrap().unwrap();
-    let checkpoint_ = state::Checkpoint {
-        parent: None,
-        hash: project_hash,
-    };
-    assert_eq!(checkpoint, checkpoint_);
-
+    let re_org: Org = client.get_org(org_id.clone()).await.unwrap().unwrap();
+    assert_eq!(re_org.members.len(), 2);
     assert!(
-        client
-            .get_project(project_name.clone(), org_id.clone())
-            .await
-            .unwrap()
-            .is_some(),
-        "Registered project not found in project list"
-    );
-
-    let org: Org = client.get_org(org_id.clone()).await.unwrap().unwrap();
-    assert!(
-        org.projects.contains(&project_name),
+        re_org.members.contains(&author_id),
         format!(
-            "Expected project id {} in Org {} with projects {:?}",
-            project_name, org_id, org.projects
+            "Expected author id {} in Org {} with members {:?}",
+            author_id, org_id, re_org.members
+        )
+    );
+    assert!(
+        re_org.members.contains(&user_id),
+        format!(
+            "Expected user id {} in Org {} with members {:?}",
+            user_id, org_id, re_org.members
         )
     );
 
@@ -123,7 +186,7 @@ async fn register_org() {
     let _ = env_logger::try_init();
     let node_host = url::Host::parse("127.0.0.1").unwrap();
     let client = Client::create_with_executor(node_host).await.unwrap();
-    let (author, _) = key_pair_with_associated_user(&client).await;
+    let (author, user_id) = key_pair_with_associated_user(&client).await;
 
     let initial_balance = client.free_balance(&author.public()).await.unwrap();
 
@@ -145,7 +208,7 @@ async fn register_org() {
     assert!(opt_org.is_some(), "Registered org not found in orgs list");
     let org = opt_org.unwrap();
     assert_eq!(org.id, register_org_message.org_id);
-    assert_eq!(org.members, vec![author.public()]);
+    assert_eq!(org.members, vec![user_id]);
     assert!(org.projects.is_empty());
 
     assert_eq!(