+++ runtime/src/registry.rs
@@ -63,6 +63,17 @@ where
 /// Funds that are credited to the block author for every block.
 pub const BLOCK_REWARD: Balance = 1000;
 
+// Placeholder data to be exported by the client so we can implement the UI in
+// Upstream.
+/// Deposit for registering a user.
+pub const REGISTER_USER_DEPOSIT: Balance = 10;
+/// Deposit for registering a project.
+pub const REGISTER_PROJECT_DEPOSIT: Balance = 10;
+/// Deposit for registering an org.
+pub const REGISTER_ORG_DEPOSIT: Balance = 10;
+/// Deposit for registering a member on an org.
+pub const REGISTER_MEMBER_DEPOSIT: Balance = 10;
+
 pub mod store {
     use super::*;
 
@@ -92,9 +103,11 @@ pub mod store {
             // We use the blake2_128_concat hasher so that the ProjectId can be extracted from the
             // key.
             pub Projects: map hasher(blake2_128_concat) ProjectId => Option<state::Project>;
+
             // The below map indexes each existing project's id to the
             // checkpoint id that it was registered with.
             pub InitialCheckpoints: map hasher(opaque_blake2_256) ProjectId => Option<CheckpointId>;
+
             // The below map indexes each checkpoint's id to the checkpoint
             // it points to, should it exist.
             pub Checkpoints: map hasher(opaque_blake2_256) CheckpointId => Option<state::Checkpoint>;
@@ -150,35 +163,63 @@ decl_module! {
         pub fn register_project(origin, message: message::RegisterProject) -> DispatchResult {
             let sender = ensure_signed(origin)?;
 
-            let org = match store::Orgs::get(message.org_id.clone()) {
-                None => return Err(RegistryError::InexistentOrg.into()),
-                Some(o) => o,
-            };
-
-            if !org.members.contains(&sender) {
-                return Err(RegistryError::InsufficientSenderPermissions.into());
-            }
-
             if store::Checkpoints::get(message.checkpoint_id).is_none() {
                 return Err(RegistryError::InexistentCheckpointId.into())
             }
 
-            let project_id = (message.project_name.clone(), message.org_id.clone());
-
+            let project_id = (message.project_name.clone(), message.project_domain.clone());
             if store::Projects::get(project_id.clone()).is_some() {
                 return Err(RegistryError::DuplicateProjectId.into());
+            };
+
+            match &message.project_domain {
+                ProjectDomain::Org(org_id) => {
+                    let org = store::Orgs::get(org_id).ok_or(RegistryError::InexistentOrg)?;
+                    if !org_has_member_with_account(&org, sender) {
+                        return Err(RegistryError::InsufficientSenderPermissions.into());
+                    }
+                    store::Orgs::insert(org_id, org.add_project(message.project_name.clone()));
+                },
+                ProjectDomain::User(user_id) => {
+                    let user = store::Users::get(user_id).ok_or(RegistryError::InexistentUser)?;
+                    if user.account_id != sender {
+                        return Err(RegistryError::InsufficientSenderPermissions.into());
+                    }
+                    store::Users::insert(user_id, user.add_project(message.project_name.clone()));
+                },
             };
 
             let new_project = state::Project {
                 current_cp: message.checkpoint_id,
                 metadata: message.metadata
             };
-
             store::Projects::insert(project_id.clone(), new_project);
-            store::Orgs::insert(message.org_id.clone(), org.add_project(message.project_name.clone()));
             store::InitialCheckpoints::insert(project_id, message.checkpoint_id);
 
-            Self::deposit_event(Event::ProjectRegistered(message.project_name, message.org_id));
+            Self::deposit_event(Event::ProjectRegistered(message.project_name, message.project_domain));
+            Ok(())
+        }
+
+        #[weight = SimpleDispatchInfo::InsecureFreeNormal]
+        pub fn register_member(origin, message: message::RegisterMember) -> DispatchResult {
+            let sender = ensure_signed(origin)?;
+
+            let org = store::Orgs::get(message.org_id.clone()).ok_or(RegistryError::InexistentOrg)?;
+            if !org_has_member_with_account(&org, sender) {
+                return Err(RegistryError::InsufficientSenderPermissions.into());
+            }
+
+            if store::Users::get(message.user_id.clone()).is_none() {
+                return Err(RegistryError::InexistentUser.into());
+            }
+
+            if org.members.contains(&message.user_id) {
+                return Err(RegistryError::AlreadyAMember.into());
+            }
+
+            let org_with_member = org.add_member(message.user_id.clone());
+            store::Orgs::insert(message.org_id.clone(), org_with_member);
+            Self::deposit_event(Event::MemberRegistered(message.user_id, message.org_id));
             Ok(())
         }
 
@@ -186,10 +227,11 @@ decl_module! {
         pub fn register_org(origin, message: message::RegisterOrg) -> DispatchResult {
             let sender = ensure_signed(origin)?;
 
-            match store::Orgs::get(message.org_id.clone()) {
-                None => {},
-                Some(_) => return Err(RegistryError::DuplicateOrgId.into()),
+            if store::Orgs::get(message.org_id.clone()).is_some() {
+                return Err(RegistryError::DuplicateOrgId.into());
             }
+
+            let user = get_user_with_account(sender).ok_or(RegistryError::AuthorHasNoAssociatedUser)?;
 
             let random_account_id = AccountId::unchecked_from(
                 pallet_randomness_collective_flip::Module::<T>::random(
@@ -199,7 +241,7 @@ decl_module! {
 
             let new_org = state::Org {
                 account_id: random_account_id,
-                members: vec![sender],
+                members: vec![user.id],
                 projects: Vec::new(),
             };
             store::Orgs::insert(message.org_id.clone(), new_org);
@@ -210,7 +252,8 @@ decl_module! {
         #[weight = SimpleDispatchInfo::InsecureFreeNormal]
         pub fn unregister_org(origin, message: message::UnregisterOrg) -> DispatchResult {
             fn can_be_unregistered(org: state::Org, sender: AccountId) -> bool {
-                org.members == vec![sender] && org.projects.is_empty()
+                org.projects.is_empty() && get_user_with_account(sender)
+                    .map(|user| org.members == vec![user.id]).unwrap_or(false)
             }
 
             let sender = ensure_signed(origin)?;
@@ -238,13 +281,8 @@ decl_module! {
                 return Err(RegistryError::DuplicateUserId.into())
             }
 
-            // TODO(xla): This is a naive first version of the check to see if an account is
-            // already associated to a user. While fine for small dataset this needs to be reworked
-            // in the future.
-            for user in store::Users::iter() {
-                if sender == user.1.account_id {
-                    return Err(RegistryError::UserAccountAssociated.into())
-                }
+            if get_user_with_account(sender).is_some() {
+                return Err(RegistryError::UserAccountAssociated.into())
             }
 
             let new_user = state::User {
@@ -258,35 +296,28 @@ decl_module! {
 
         #[weight = SimpleDispatchInfo::InsecureFreeNormal]
         pub fn unregister_user(origin, message: message::UnregisterUser) -> DispatchResult {
-            fn can_be_unregistered(user: state::User, sender: AccountId) -> bool {
-                user.account_id == sender && user.projects.is_empty()
-            }
-
             let sender = ensure_signed(origin)?;
+            let sender_user = get_user_with_account(sender).ok_or(RegistryError::InexistentUser)?;
 
-            match store::Users::get(message.user_id.clone()) {
-                None => Err(RegistryError::InexistentUser.into()),
-                Some(user) => {
-                    if can_be_unregistered(user, sender) {
-                        store::Users::remove(message.user_id.clone());
-                        Self::deposit_event(Event::UserUnregistered(message.user_id));
-                        Ok(())
-                    }
-                    else {
-                        Err(RegistryError::UnregisterableUser.into())
-                    }
-                }
+            if sender_user.id != message.user_id {
+                return Err(RegistryError::InsufficientSenderPermissions.into());
             }
+            if find_org(|org| org.members.contains(&sender_user.id)).is_some() {
+                return Err(RegistryError::UnregisterableUser.into());
+            }
+
+            store::Users::remove(message.user_id.clone());
+            Self::deposit_event(Event::UserUnregistered(message.user_id));
+            Ok(())
         }
 
         #[weight = SimpleDispatchInfo::InsecureFreeNormal]
         pub fn transfer_from_org(origin, message: message::TransferFromOrg) -> DispatchResult {
             let sender = ensure_signed(origin)?;
-            let org = match store::Orgs::get(message.org_id) {
-                None => return Err(RegistryError::InexistentOrg.into()),
-                Some(o) => o,
-            };
-            if org.members.contains(&sender) {
+            let org = store::Orgs::get(message.org_id)
+                .ok_or(RegistryError::InexistentOrg)?;
+
+            if org_has_member_with_account(&org, sender) {
                 <crate::Balances as Currency<_>>::transfer(
                     &org.account_id,
                     &message.recipient,
@@ -336,12 +367,17 @@ decl_module! {
             if store::Checkpoints::get(message.new_checkpoint_id).is_none() {
                 return Err(RegistryError::InexistentCheckpointId.into())
             }
-            let project_id = (message.project_name.clone(), message.org_id.clone());
+            let project_id = (message.project_name.clone(), message.project_domain.clone());
             let opt_project = store::Projects::get(project_id.clone());
-            let opt_org = store::Orgs::get(message.org_id.clone());
+
+            let org_id = match &message.project_domain {
+                ProjectDomain::Org(org_id) => org_id,
+                ProjectDomain::User(_) => panic!("TODO(nuno"),
+            };
+            let opt_org = store::Orgs::get(org_id.clone());
             let new_project = match (opt_project, opt_org) {
                 (Some(prj), Some(org)) => {
-                    if !org.members.contains(&sender) {
+                    if !org_has_member_with_account(&org, sender) {
                         return Err(RegistryError::InsufficientSenderPermissions.into())
                     }
                     state::Project {
@@ -365,7 +401,7 @@ decl_module! {
 
             Self::deposit_event(Event::CheckpointSet(
                 message.project_name.clone(),
-                message.org_id.clone(),
+                message.project_domain.clone(),
                 message.new_checkpoint_id
             ));
             Ok(())
@@ -399,13 +435,40 @@ decl_module! {
 
     }
 }
+
+// TODO(xla): This is a naive first version of the check to see if an account is
+// already associated to a user. While fine for small dataset this needs to be reworked
+// in the future.
+pub fn get_user_with_account(account_id: AccountId) -> Option<User> {
+    store::Users::iter()
+        .find(|(_, user)| user.account_id == account_id)
+        .map(|(id, user)| User::new(id, user))
+}
+
+pub fn find_org(predicate: impl Fn(&state::Org) -> bool) -> Option<state::Org> {
+    store::Orgs::iter()
+        .find(|(_, org)| predicate(org))
+        .map(|(_, org)| org)
+}
+
+/// Check whether the user associated with the given account_id is a member of the given org.
+/// Return false if the account doesn't have an associated user or if said user is not a member
+/// of the org.
+pub fn org_has_member_with_account(org: &state::Org, account_id: AccountId) -> bool {
+    match get_user_with_account(account_id) {
+        Some(user) => org.members.contains(&user.id),
+        None => false,
+    }
+}
+
 decl_event!(
     pub enum Event {
         CheckpointCreated(CheckpointId),
-        CheckpointSet(ProjectName, Id, CheckpointId),
+        CheckpointSet(ProjectName, ProjectDomain, CheckpointId),
+        MemberRegistered(Id, Id),
         OrgRegistered(Id),
         OrgUnregistered(Id),
-        ProjectRegistered(ProjectName, Id),
+        ProjectRegistered(ProjectName, ProjectDomain),
         UserRegistered(Id),
         UserUnregistered(Id),
     }
@@ -485,7 +548,7 @@ mod test {
     fn projects_decode_key_identity() {
         let org_id = Id::try_from("monadic").unwrap();
         let project_name = ProjectName::try_from("radicle".to_string()).unwrap();
-        let project_id: ProjectId = (project_name, org_id);
+        let project_id: ProjectId = (project_name, ProjectDomain::Org(org_id));
         let hashed_key = store::Projects::storage_map_final_key(project_id.clone());
         let decoded_key = store::Projects::decode_key(&hashed_key).unwrap();
         assert_eq!(decoded_key, project_id);